refactor: split context providers from hooks to fix react-refresh warning
Moved useAuth/AuthContext and useTheme/ThemeContext into sibling
useAuth.ts/useTheme.ts modules, leaving AuthContext.tsx and
ThemeContext.tsx exporting only their Provider components. This
satisfies eslint-plugin-react-refresh/only-export-components and lets
HMR reload the providers cleanly. Updated all import sites and
vi.mock paths in tests; behavior is unchanged.
