+++ client/src/context/useAuth.ts
@@ -1,0 +1,27 @@
+import { createContext, useContext } from 'react';
+
+export interface AuthState {
+  token: string | null;
+  refreshToken: string | null;
+  userId: number | null;
+  username: string | null;
+  role: string | null;
+  mustChangePassword: boolean;
+}
+
+export interface AuthContextType extends AuthState {
+  login: (email: string, password: string) => Promise<void>;
+  register: (email: string, password: string) => Promise<void>;
+  logout: () => void;
+  clearMustChangePassword: () => void;
+  isAuthenticated: boolean;
+  isAdmin: boolean;
+}
+
+export const AuthContext = createContext<AuthContextType | null>(null);
+
+export function useAuth() {
+  const ctx = useContext(AuthContext);
+  if (!ctx) throw new Error('useAuth must be used inside AuthProvider');
+  return ctx;
+}
