+++ client/src/pages/ChangePasswordPage.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "@/context/AuthContext";
+import { useAuth } from "@/context/useAuth";
 import userService from "@/services/userService";
 import Button from "@/components/common/Button";
