+++ client/src/pages/LoginPage.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { NavLink, useNavigate } from "react-router-dom";
-import { useAuth } from "@/context/AuthContext";
+import { useAuth } from "@/context/useAuth";
 import Button from "@/components/common/Button";
 
 const LoginPage = () => {
