+++ client/test/context/AuthContext.test.tsx
@@ -1,6 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { render, screen, act, fireEvent } from '@testing-library/react'
-import { AuthProvider, useAuth } from '@/context/AuthContext'
+import { AuthProvider } from '@/context/AuthContext'
+import { useAuth } from '@/context/useAuth'
 
 vi.mock('@/services/userService', () => ({
   default: {
