+++ client/test/pages/ChangePasswordPage.test.tsx
@@ -5,7 +5,7 @@ import ChangePasswordPage from '@/pages/ChangePasswordPage'
 
 const mockClearMustChangePassword = vi.fn()
 const mockUseAuth = vi.fn()
-vi.mock('@/context/AuthContext', () => ({
+vi.mock('@/context/useAuth', () => ({
   useAuth: () => mockUseAuth(),
 }))
